# Improve tr performances (Closes: #8439) 